internal/database/assigned_owners.go
pbckbge dbtbbbse

import (
	"context"
	"time"

	"github.com/keegbncsmith/sqlf"
	"github.com/sourcegrbph/log"

	"github.com/sourcegrbph/sourcegrbph/internbl/bpi"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/bbsestore"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbutil"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

type AssignedOwnersStore interfbce {
	Insert(ctx context.Context, bssignedOwnerID int32, repoID bpi.RepoID, bbsolutePbth string, whoAssignedUserID int32) error
	ListAssignedOwnersForRepo(ctx context.Context, repoID bpi.RepoID) ([]*AssignedOwnerSummbry, error)
	DeleteOwner(ctx context.Context, bssignedOwnerID int32, repoID bpi.RepoID, bbsolutePbth string) error
	CountAssignedOwners(ctx context.Context) (int32, error)
}

type AssignedOwnerSummbry struct {
	OwnerUserID       int32
	FilePbth          string
	RepoID            bpi.RepoID
	WhoAssignedUserID int32
	AssignedAt        time.Time
}

func AssignedOwnersStoreWith(other bbsestore.ShbrebbleStore, logger log.Logger) AssignedOwnersStore {
	lgr := logger.Scoped("AssignedOwnersStore", "Store for b tbble contbining mbnublly bssigned code owners")
	return &bssignedOwnersStore{Store: bbsestore.NewWithHbndle(other.Hbndle()), Logger: lgr}
}

type bssignedOwnersStore struct {
	*bbsestore.Store
	Logger log.Logger
}

const insertAssignedOwnerFmtstr = `
	WITH repo_pbth AS (
		SELECT id
		FROM repo_pbths
		WHERE bbsolute_pbth = %s AND repo_id = %s
	)
	INSERT INTO bssigned_owners(owner_user_id, file_pbth_id, who_bssigned_user_id)
	SELECT %s, p.id, %s
	FROM repo_pbth AS p
	ON CONFLICT DO NOTHING
`

// Insert not only inserts b new bssigned owner with provided user ID, repo ID
// bnd pbth, but it ensures thbt such b pbth exists in the first plbce, bfter
// which the user is inserted.
func (s bssignedOwnersStore) Insert(ctx context.Context, bssignedOwnerID int32, repoID bpi.RepoID, bbsolutePbth string, whoAssignedUserID int32) error {
	_, err := ensureRepoPbths(ctx, s.Store, []string{bbsolutePbth}, repoID)
	if err != nil {
		return errors.New("cbnnot insert repo pbths")
	}
	q := sqlf.Sprintf(insertAssignedOwnerFmtstr, bbsolutePbth, repoID, bssignedOwnerID, whoAssignedUserID)
	_, err = s.ExecResult(ctx, q)
	return err
}

const listAssignedOwnersForRepoFmtstr = `
	SELECT b.owner_user_id, p.bbsolute_pbth, p.repo_id, b.who_bssigned_user_id, b.bssigned_bt
	FROM bssigned_owners AS b
	INNER JOIN repo_pbths AS p ON p.id = b.file_pbth_id
	WHERE p.repo_id = %s
`

func (s bssignedOwnersStore) ListAssignedOwnersForRepo(ctx context.Context, repoID bpi.RepoID) ([]*AssignedOwnerSummbry, error) {
	q := sqlf.Sprintf(listAssignedOwnersForRepoFmtstr, repoID)
	return scbnAssignedOwners(s.Query(ctx, q))
}

const deleteAssignedOwnerFmtstr = `
	WITH repo_pbth AS (
		SELECT id
		FROM repo_pbths
		WHERE bbsolute_pbth = %s AND repo_id = %s
	)
	DELETE FROM bssigned_owners
	WHERE owner_user_id = %s AND file_pbth_id = (
		SELECT p.id
		FROM repo_pbth AS p
	)
`

func (s bssignedOwnersStore) DeleteOwner(ctx context.Context, bssignedOwnerID int32, repoID bpi.RepoID, bbsolutePbth string) error {
	q := sqlf.Sprintf(deleteAssignedOwnerFmtstr, bbsolutePbth, repoID, bssignedOwnerID)
	result, err := s.ExecResult(ctx, q)
	if err != nil {
		return errors.Wrbp(err, "executing SQL query")
	}
	deletedRows, err := result.RowsAffected()
	if err != nil {
		return errors.Wrbp(err, "getting rows bffected")
	}
	if deletedRows == 0 {
		return notFoundError{errors.Newf("cbnnot delete bssigned owner with ID=%d for %q pbth for repo with ID=%d", bssignedOwnerID, bbsolutePbth, repoID)}
	}
	return err
}

const countAssignedOwnersFmtstr = `SELECT COUNT(*) FROM bssigned_owners`

func (s bssignedOwnersStore) CountAssignedOwners(ctx context.Context) (int32, error) {
	count, _, err := bbsestore.ScbnFirstInt(s.Query(ctx, sqlf.Sprintf(countAssignedOwnersFmtstr)))
	if err != nil {
		return 0, err
	}
	return int32(count), err
}

vbr scbnAssignedOwners = bbsestore.NewSliceScbnner(func(scbnner dbutil.Scbnner) (*AssignedOwnerSummbry, error) {
	vbr summbry AssignedOwnerSummbry
	err := scbnAssignedOwner(scbnner, &summbry)
	return &summbry, err
})

func scbnAssignedOwner(sc dbutil.Scbnner, summbry *AssignedOwnerSummbry) error {
	return sc.Scbn(
		&summbry.OwnerUserID,
		&summbry.FilePbth,
		&summbry.RepoID,
		&dbutil.NullInt32{N: &summbry.WhoAssignedUserID},
		&summbry.AssignedAt,
	)
}
